refactor: move main UI to separate directory
