+++ lib/welcome_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_widgets/screens/home_screen.dart';
 import 'package:flutter_widgets/widgets/buttons/gradient_button/MyGradientButton.dart';
 import 'package:flutter_widgets/widgets/cupertinoActionSheet/CupertinoActionSheet.dart';
 import 'package:flutter_widgets/widgets/drawer/Drawer.dart';
@@ -17,16 +18,19 @@ class WelcomeScreen extends StatefulWidget {
 
 class _WelcomeScreenState extends State<WelcomeScreen> {
   int _currentIndex = 0;
-  static const List<Widget> _widgetOptions = <Widget>[
-    Text(
-      'Index 0: Home',
+  static List<Widget> _widgetOptions = <Widget>[
+    HomeScreen(),
+    Padding(
+      padding: const EdgeInsets.all(10.0),
+      child: Center(
+        child: Text(
+          'Add some widget to display here!',
+          style: TextStyle(fontSize: 20.0),
+          textAlign: TextAlign.center,
+        ),
+      ),
     ),
-    Text(
-      'Index 1: Business',
-    ),
-    Text(
-      'Index 2: School',
-    ),
+    Error404(),
   ];
   @override
   Widget build(BuildContext context) {
@@ -36,61 +40,7 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
           "Awesome flutter widgets",
         ),
       ),
-      body: Container(
-        width: MediaQuery.of(context).size.width,
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: [
-            MyGradientButton(
-              key: Key('actionSheetKey'),
-              title: 'Show Action Sheet',
-              onPress: () async {
-                await showCupertinoModalPopup(
-                    context: context,
-                    builder: (context) => MyCupertinoActionSheet());
-              },
-            ),
-            MyGradientButton(
-              title: 'Show Expansion Tile',
-              onPress: () {
-                Navigator.pushNamed(context, MyExpansionTile.id);
-              },
-            ),
-            MyGradientButton(
-              title: 'Swipe Up Panel',
-              onPress: () {
-                Navigator.pushNamed(context, SwipeUpPanel.id);
-              },
-            ),
-            MyGradientButton(
-              title: 'Alert Dialog',
-              onPress: () {
-                var dialog = CustomAlertDialog(
-                    title: "View Swipe panel",
-                    message:
-                        "Are you sure, do you want to view swipe up panel?",
-                    onPositivePressed: () {
-                      Navigator.popAndPushNamed(context, SwipeUpPanel.id);
-                    },
-                    positiveBtnText: 'Yes',
-                    negativeBtnText: 'No',
-                    bgColor: Colors.white);
-
-                showDialog(
-                    context: context,
-                    builder: (BuildContext context) => dialog);
-              },
-            ),
-            MyGradientButton(
-              title: '404 Error page',
-              onPress: () {
-                Navigator.pushNamed(context, Error404.id);
-              },
-            ),
-          ],
-        ),
-      ),
+      body: _widgetOptions.elementAt(_currentIndex),
       drawer: UserAccountDrawer(),
       bottomNavigationBar: BottomNavigationBar(
         currentIndex: _currentIndex,
